Annotate and refactor RX observables in OfflineBaseMapViewer

+++ gnd/src/main/java/com/google/android/gnd/ui/offlinebasemap/viewer/OfflineBaseMapViewerViewModel.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 Google LLC
+ * Copyright 2021 Google LLC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -24,14 +24,19 @@ import androidx.lifecycle.LiveDataReactiveStreams;
 import com.google.android.gnd.model.basemap.OfflineBaseMap;
 import com.google.android.gnd.model.basemap.tile.TileSource;
 import com.google.android.gnd.repository.OfflineBaseMapRepository;
+import com.google.android.gnd.rx.Nil;
+import com.google.android.gnd.rx.annotations.Hot;
 import com.google.android.gnd.ui.common.AbstractViewModel;
 import com.google.common.collect.ImmutableSet;
 import dagger.hilt.android.qualifiers.ApplicationContext;
 import io.reactivex.Completable;
+import io.reactivex.Flowable;
 import io.reactivex.processors.BehaviorProcessor;
+import io.reactivex.processors.PublishProcessor;
 import java.io.File;
 import java.lang.ref.WeakReference;
 import javax.inject.Inject;
+import timber.log.Timber;
 
 /**
  * View model for the OfflineAreaViewerFragment. Manges offline area deletions and calculates the
@@ -39,45 +44,41 @@ import javax.inject.Inject;
  */
 public class OfflineBaseMapViewerViewModel extends AbstractViewModel {
 
-  private final BehaviorProcessor<OfflineBaseMapViewerFragmentArgs> argsProcessor;
+  @Hot private final BehaviorProcessor<OfflineBaseMapViewerFragmentArgs> argsProcessor;
+  @Hot private final PublishProcessor<Nil> removeClickProcessor;
   private final OfflineBaseMapRepository offlineBaseMapRepository;
   private final WeakReference<Context> context;
   public LiveData<Double> areaStorageSize;
   public LiveData<String> areaName;
-  private LiveData<OfflineBaseMap> offlineArea;
+  private final LiveData<OfflineBaseMap> offlineArea;
 
   @Inject
   public OfflineBaseMapViewerViewModel(
       OfflineBaseMapRepository offlineBaseMapRepository, @ApplicationContext Context context) {
     this.argsProcessor = BehaviorProcessor.create();
+    this.removeClickProcessor = PublishProcessor.create();
     this.offlineBaseMapRepository = offlineBaseMapRepository;
     this.context = new WeakReference<>(context);
+    @Hot
+    Flowable<OfflineBaseMap> _offlineArea =
+        this.argsProcessor.switchMap(
+            args ->
+                this.offlineBaseMapRepository
+                    .getOfflineArea(args.getOfflineAreaId())
+                    .toFlowable()
+                    .doOnError(
+                        throwable ->
+                            Timber.e(
+                                throwable, "Couldn't render area: %s", args.getOfflineAreaId())));
     this.areaName =
-        LiveDataReactiveStreams.fromPublisher(
-            this.argsProcessor.switchMap(
-                args ->
-                    this.offlineBaseMapRepository
-                        .getOfflineArea(args.getOfflineAreaId())
-                        .toFlowable()
-                        .map(OfflineBaseMap::getName)));
+        LiveDataReactiveStreams.fromPublisher(_offlineArea.map(OfflineBaseMap::getName));
     this.areaStorageSize =
         LiveDataReactiveStreams.fromPublisher(
-            this.argsProcessor.switchMap(
-                args ->
-                    this.offlineBaseMapRepository
-                        .getOfflineArea(args.getOfflineAreaId())
-                        .toFlowable()
-                        .flatMap(
-                            offlineBaseMapRepository
-                                ::getIntersectingDownloadedTileSourcesOnceAndStream)
-                        .map(this::tileSourcesToTotalStorageSize)));
-    this.offlineArea =
-        LiveDataReactiveStreams.fromPublisher(
-            this.argsProcessor.switchMap(
-                args ->
-                    this.offlineBaseMapRepository
-                        .getOfflineArea(args.getOfflineAreaId())
-                        .toFlowable()));
+            _offlineArea
+                .flatMap(
+                    offlineBaseMapRepository::getIntersectingDownloadedTileSourcesOnceAndStream)
+                .map(this::tileSourcesToTotalStorageSize));
+    this.offlineArea = LiveDataReactiveStreams.fromPublisher(_offlineArea);
   }
 
   private Double tileSourcesToTotalStorageSize(ImmutableSet<TileSource> tileSources) {
@@ -97,9 +98,20 @@ public class OfflineBaseMapViewerViewModel extends AbstractViewModel {
   /**
    * Removes the offline area associated with this viewmodel from the device by removing all tile
    * sources that are not included in other areas and removing the area from the db.
+   *
+   * Terminates the upstream click processor so that the resulting stream completes.
    */
-  public Completable onRemoveClick() {
-    return offlineBaseMapRepository.deleteArea(this.offlineArea.getValue().getId());
+  @Hot(terminates = true)
+  public Completable onRemoveArea() {
+    return this.removeClickProcessor.flatMapCompletable(
+        __ -> {
+          if (this.offlineArea.getValue() == null) {
+            return Completable.error(new Throwable("Could not remove nonexistent area."));
+          }
+
+          this.removeClickProcessor.onComplete();
+          return offlineBaseMapRepository.deleteArea(this.offlineArea.getValue().getId());
+        });
   }
 
   /** Returns the offline area associated with this view model. */
@@ -110,5 +122,11 @@ public class OfflineBaseMapViewerViewModel extends AbstractViewModel {
   /** Gets a single offline area by the id passed to the OfflineAreaViewerFragment's arguments. */
   public void loadOfflineArea(OfflineBaseMapViewerFragmentArgs args) {
     this.argsProcessor.onNext(args);
+  }
+
+  /** Deletes the area associated with this viewmodel. */
+  public void removeArea() {
+    Timber.d("Removing offline area %s", this.offlineArea.getValue());
+    this.removeClickProcessor.onNext(Nil.NIL);
   }
 }
